+++ src/renderer/models/Note.js
@@ -1,0 +1,7 @@
+import Resource from './Resource'
+// -----------------------------------------------------------------------------
+class Note extends Resource {
+    static RESOURCE_NAME = 'notes'
+}
+// -----------------------------------------------------------------------------
+export default Note